// chromato-app/src/Store/Enums/Endpoints.ts
export enum Endpoints {
  InitPalette = "palette/initialize",
  CreatePalette = "palette/create",
  DeletePalette = "palette/delete",
  AddPaletteColor = "palette/color/add",
  RemovePaletteColor = "palette/color/remove",
  SetPaletteColor = "palette/color/set",
  SuggestPaletteColors = "palette/color/suggest",
  SetBackgroundColor = "active/color/background",
  SetForegroundColor = "active/color/foreground",
  SetColor = "active/color/auto",
  SetColor2D = "palette/active/color2D",
  SetCombinationStrategy = "palette/combinationStrategy/set",
  SetReadabilityPolicy = "palette/readabilityPolicy/set",
  SetReadabilityDisplayStrategy = "palette/readabilityDisplayStrategy/set",
  SetGenerationStrategy = "palette/generationStrategy/set",
  SetColorBehavior = "palette/colorBehavior/set",  
  SetSuggestionStrategy = "palette/suggestionStrategy/set",
  SetTemplate = "palette/template/set",
  SetSuggestionSize = "palette/suggestionSize/set",
  SetElementSize = "palette/elementSize/set",
}